+++ src/core/modules/atomic/crypto/__init__.py
@@ -40,4 +40,9 @@ try:
 except ImportError:
     pass
 
+try:
+    from .totp import *
+except ImportError:
+    pass
+
 __all__ = []